--- src/types/appConfig.ts
@@ -1,7 +1,0 @@
-import { DeploymentItemConfig, HookConfig } from '../container/types';
-
-export interface AppConfig {
-  app: string;
-  deployments: { [name: string]: DeploymentItemConfig };
-  hooks: HookConfig[];
-}
