--- src/types/index.ts
@@ -1,6 +1,0 @@
-export * from './app';
-export * from './appConfig';
-export * from './cliContext';
-export * from './clusterConfig';
-export * from './globalConfig';
-export * from './tenantConfig';
