--- src/types/tenantConfig.ts
@@ -1,7 +1,0 @@
-export interface TenantConfig {
-  version: string;
-  user_config: {
-    api_key: string;
-    master_key: string;
-  };
-}
